- Converted all base type to their matching uintxx_t type - Created the setTempo() method - Fixed META_EVENT tempo handling
git-svn-id: http://localhost/svn/libMidi/trunk@15 28eb8d37-f890-1e4f-99d2-a32480d331fc

+++ libMidi/varlength.h
@@ -20,7 +20,7 @@ inline size_t getVariableLengthMask(int iIndex)
 }
 
 template <typename T>
-size_t fwriteVariableLength(const T & iValue, int iMinOutputSize, FILE * f)
+size_t fwriteVariableLength(const T & iValue, size_t iMinOutputSize, FILE * f)
 {
   size_t writeSize = 0;
